🐛 fix: a value on its way into a string converts the way C# converts it
"a" + name with a null name wrote "anull" in the browser and "a" on the
server; "v=" + flag wrote "v=true" here and "v=True" there; $"{n}" with
a null int? wrote "null". Text a user reads, wrong in one of the two
places the same C# runs, and nothing to say so.

StringConversion decides it once, for concatenation and interpolation
alike: a null is the empty string, a bool is True/False (the runtime's
format helper now spells it that way), a nullable value follows its
value or is nothing, an enum is its member name, and a string that may
be null — annotated `string?`, or a string from code with no nullable
context — reads as itself or as nothing. A `string` under nullable-
enabled code is left bare, so the component pins gain no noise. The
nullable-lifting branch yields to concatenation, which used to capture
"v=" + n and lift it as arithmetic.

Four more of the conversion gaps close beside it: a compound `/=` on
integers divides as integers (built as IR — a ternary on the right is
fenced by the writer, not by a template), a float literal single
precision actually changes is written as Math.fround(…) so (double)0.1f
> 0.1 holds and 0.1f == 0.1 does not, Math.Round on a decimal rounds
the decimal half-to-even (the twin grew round(digits)), and Equals on a
bare object compares boxed values by value.

The conversion baseline goes from 21 divergences to 11 — what is left
is fixed-width integer arithmetic and checked contexts, float results,
char increments and enum arithmetic: the next slice. Sixteen component
modules change, every edit a Math.fround around a float literal that
was never the double it was written as.

+++ src/eQuantic.UI.Runtime/src/shared/components/Avatar.ts
@@ -35,14 +35,14 @@ export class Avatar extends StatelessComponent {
         let clipped = this.initials.length > 2 ? this.initials.substring(0, 0 + 2) : this.initials;
         let hasInitials = clipped.length > 0;
         let glyphSize = (() => { const _s = this.size; if (_s === 'small') return 16; if (_s === 'medium') return 20; if (_s === 'large') return 24; return 32; })();
-        let face = hasInitials ? new Text(clipped, 'caption', tint.onSubtle, 1, 'start', false, false, null, { styleOverride: new TypeStyle(labelSize, labelSize, 'semiBold', 0, 1.3) }) : new Icon('person', glyphSize, theme.textMuted);
+        let face = hasInitials ? new Text(clipped, 'caption', tint.onSubtle, 1, 'start', false, false, null, { styleOverride: new TypeStyle(labelSize, labelSize, 'semiBold', 0, Math.fround(1.3)) }) : new Icon('person', glyphSize, theme.textMuted);
         let content = face.centered();
         let circle = new Box(new BoxStyle({ width: side, height: side, background: hasInitials ? tint.subtle : theme.surfaceSubtle, cornerRadius: new CornerRadii(theme.shape('full')) }), content);
         return this.status === 'none' ? circle : this.withStatusDot(circle, side, theme);
     }
 
     withStatusDot(face: VisualNode, side: number, theme: any) {
-        let dotSide = $eq.math.round(side / 3.3);
+        let dotSide = $eq.math.round(side / Math.fround(3.3));
         let dotFill = this.status === 'online' ? theme.colors('success').base : theme.textMuted;
         let dot = new Box(new BoxStyle({ width: dotSide, height: dotSide, background: dotFill, cornerRadius: new CornerRadii(theme.shape('full')), borderWidth: 2, borderColor: theme.surface }));
         let stack = new Stack();
